gui: fix receive view and release shasums
